+++ utils/ExampleRunner/example-info.json
@@ -250,6 +250,10 @@
         "name": "Crosshairs",
         "description": "Here we demonstrate crosshairs linking three orthogonal views of the same data"
       },
+      "overlayGrid": {
+        "name": "Overlay Grid",
+        "description": "Demonstrate overlay grid tool usage with three viewports one for each orientation"
+      },
       "referenceLines": {
         "name": "Reference Lines",
         "description": "Demonstrate reference line tool for rendering viewports location with respect to each other"
